src/dimensional_awareness.py
from qiskit import QuantumCircuit, transpile
from qiskit_aer import AerSimulator
from qiskit.visualization import plot_histogram
import numpy as np
from model_tuning import tune_logistic_regression

def advanced_quantum_circuit():
    qc = QuantumCircuit(3)
    qc.h([0, 1, 2])  # Apply Hadamard gate to all three qubits
    qc.cx(0, 1)      # Apply CNOT gate between qubit 0 and qubit 1
    qc.cx(1, 2)      # Apply CNOT gate between qubit 1 and qubit 2
    qc.ccx(0, 1, 2)  # Apply Toffoli gate with qubits 0, 1 as control and qubit 2 as target
    qc.measure_all() # Measure all qubits

    simulator = AerSimulator()  # Use the Aer simulator
    qobj = transpile(qc, simulator)  # Transpile the circuit for the simulator
    result = simulator.run(qobj).result() # Run the circuit on the simulator
    counts = result.get_counts() # Get the measurement counts
    plot_histogram(counts)       # Plot a histogram of the results
    return counts

def prepare_ml_data(counts):
    data = []
    labels = []
    for state, count in counts.items():
        features = [int(bit) for bit in state] # Convert state string to list of integers
        label = 1 if count > 50 else 0         # Label based on count threshold
        data.append(features)
        labels.append(label)
    return np.array(data), np.array(labels)

def train_ml_model(data, labels):
    best_model = tune_logistic_regression(data, labels)
    return best_model
